decouple from card component
